tests/tests_requirements.txt
pytest 
pytest-cov
pytest-flask
httpx
faker
requests
mysql-connector-python
python-dotenv

